Move Particle physics in it's own folder and create rigidbody class
