+++ Core/include/Voxymore/Physics/Particles/Collisions/ParticleContactResolver.hpp
@@ -4,9 +4,9 @@
 
 #pragma once
 
+#include "ParticleContact.hpp"
 #include "Voxymore/Core/Math.hpp"
 #include "Voxymore/Core/TimeStep.hpp"
-#include "Voxymore/Physics/Collisions/ParticleContact.hpp"
 
 namespace Voxymore::Core
 {
