+++ Core/src/Physics/Particles/Components/AnchoredSpringComponent.cpp
