+++ Core/src/Physics/Particles/Particle.cpp
@@ -2,7 +2,7 @@
 // Created by ianpo on 05/01/2024.
 //
 
-#include "Voxymore/Physics/Particle.hpp"
+#include "Voxymore/Physics/Particles/Particle.hpp"
 
 namespace Voxymore::Core
 {
